017_use_hevea.diff from Debian.
This adds an html target for the doc, which is useful.
Note: The Debian description of this patch mentions they used hevea
instead of latex2html (which they used in older versions) due to
latex2html being non-free. This is however no longer the case. Still
there is nothing wrong with using hevea (also free and actively maintained),
and it creates somewhat compacter HTML docs by default, so stick with it.

Description: Use hevea instead of latex2html, now that latter has moved to
 non-free.
 0.5.19+beta20030413-2
Author: Chris Hanson <cph@debian.org>
Bug-Debian: https://bugs.debian.org/221348