Debian patch 028_suspicious_file.patch
This patch adds output of the file name when decoding split files,
the intention is to allow seeing which file caused a warning.

Description from Debian:

Add file open information so that if S_DATA_SUSPICIOUS message.occurs,
the name of the bad file is printed. There is probablly a better method
but I did not see how to obtain the filename from within the function.
Author: Peter Muir <iyhi@yahoo.com>
Bug-Debian: http://bugs.debian.org/166077